+++ app/src/main/java/com/kidemma/home_admin/tabs/kids/presentation/KidsTabScreen.kt
@@ -1,35 +1,16 @@
 package com.kidemma.home_admin.tabs.kids.presentation
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ViewList
-import androidx.compose.material.icons.filled.FilterList
-import androidx.compose.material.icons.filled.GridView
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.kidemma.R
 import com.kidemma.common.components.KidemmaLoadingOverlay
 import com.kidemma.common.ui.theme.KidemmaColors
-import com.kidemma.common.ui.theme.KidemmaDimens
 import com.kidemma.common.ui.theme.KidemmaTheme
 import com.kidemma.home_admin.tabs.kids.data.KidsTabMockProvider
 import com.kidemma.home_admin.tabs.kids.presentation.KidsTabContract.Intent.OnApplyFilter
@@ -39,6 +20,7 @@ import com.kidemma.home_admin.tabs.kids.presentation.components.KidsTabFilterDia
 import com.kidemma.home_admin.tabs.kids.presentation.ui.EmptyKidsContent
 import com.kidemma.home_admin.tabs.kids.presentation.ui.KidsGridView
 import com.kidemma.home_admin.tabs.kids.presentation.ui.KidsListView
+import com.kidemma.home_admin.tabs.kids.presentation.ui.KidsTabHeader
 import org.koin.androidx.compose.koinViewModel
 
 /*
@@ -47,7 +29,7 @@ import org.koin.androidx.compose.koinViewModel
  *
  * Created by: José Manuel Carrillo Torres
  * Created on: 09/03/26
- * Last modified: 09/03/26
+ * Last modified: 14/05/26
  */
 
 @Composable
@@ -73,16 +55,18 @@ private fun KidsTabContent(
                 .fillMaxSize()
                 .background(KidemmaColors.Background),
         ) {
-            // TODO: Add Laura's filter field
-            // KidsTabHeaderComponents()
-
-            KidsTabActions(
+            KidsTabHeader(
+                searchQuery = uiState.searchQuery,
+                onSearchQueryChange = { query ->
+                    onIntent(KidsTabContract.Intent.OnSearchQueryChange(query))
+                },
+                onClickFilter = { onIntent(OnOpenFilterDialog) },
                 isGridView = uiState.isGridView,
-                onOpenFilter = { onIntent(OnOpenFilterDialog) },
                 onToggleGridView = { onIntent(KidsTabContract.Intent.OnToggleGridView) }
             )
 
             KidsTabMainContent(
+                modifier = Modifier.weight(1f),
                 isGridView = uiState.isGridView,
                 contentState = uiState.content,
                 onKidClick = { kidId ->
@@ -102,33 +86,6 @@ private fun KidsTabContent(
             onApplyFilter = { filter ->
                 onIntent(OnApplyFilter(filter))
             },
-        )
-    }
-}
-
-@Composable
-private fun KidsTabActions(
-    isGridView: Boolean,
-    onOpenFilter: () -> Unit,
-    onToggleGridView: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(
-                top = KidsTabScreenUiConstants.ActionsTopPadding,
-                start = KidsTabScreenUiConstants.HorizontalPadding,
-                end = KidsTabScreenUiConstants.HorizontalPadding
-            ),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        FilterButton(onClick = onOpenFilter)
-
-        ToggleGridViewButton(
-            isGridView = isGridView,
-            onToggle = onToggleGridView,
         )
     }
 }
@@ -169,71 +126,10 @@ private fun KidsTabMainContent(
     }
 }
 
-@Composable
-private fun ToggleGridViewButton(
-    isGridView: Boolean,
-    onToggle: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val icon = if (isGridView) Icons.AutoMirrored.Filled.ViewList else Icons.Filled.GridView
-    val iconContentDescription = if (isGridView) stringResource(R.string.kids_list_view_content_description)
-    else stringResource(R.string.kids_grid_view_content_description)
-
-    Card(
-        onClick = onToggle,
-        modifier = modifier
-            .size(KidsTabScreenUiConstants.ButtonSize)
-            .clip(CircleShape),
-        shape = CircleShape,
-        colors = CardDefaults.cardColors(containerColor = KidemmaColors.Card),
-        elevation = CardDefaults.cardElevation(defaultElevation = KidsTabScreenUiConstants.ButtonElevation),
-    ) {
-        Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-            Icon(
-                imageVector = icon,
-                contentDescription = iconContentDescription,
-                tint = KidemmaColors.Icon,
-                modifier = Modifier.size(KidemmaDimens.IconSizeSmall),
-            )
-        }
-    }
-}
-
-@Composable
-private fun FilterButton(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Card(
-        onClick = onClick,
-        modifier = modifier
-            .size(KidsTabScreenUiConstants.ButtonSize)
-            .clip(CircleShape),
-        shape = CircleShape,
-        colors = CardDefaults.cardColors(containerColor = KidemmaColors.Card),
-        elevation = CardDefaults.cardElevation(defaultElevation = KidsTabScreenUiConstants.ButtonElevation),
-    ) {
-        Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-            Icon(
-                imageVector = Icons.Filled.FilterList,
-                contentDescription = stringResource(R.string.kids_open_filter_dialog_content_description),
-                tint = KidemmaColors.Icon,
-                modifier = Modifier.size(KidemmaDimens.IconSizeSmall),
-            )
-        }
-    }
-}
-
-private object KidsTabScreenUiConstants {
-    val HorizontalPadding = 16.dp
-    val ActionsTopPadding = 16.dp
-    val ButtonSize = 36.dp
-    val ButtonElevation = 6.dp
-}
-
 @Preview(showBackground = true)
 @Composable
-private fun KidsTabScreenPreview() {
+private fun
+        KidsTabScreenPreview() {
     KidemmaTheme {
         KidsTabContent(
             uiState = KidsTabContract.State(
